# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Shopee - OpenAPI 3.0
  description: ''
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.11
...
servers:
  - url: 'http://localhost:3000/api/v1'
...
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - auth
      summary: Refresh Token
      description: Refresh Token
      operationId: refresh_token
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/paths/~1shops/get/parameters/0'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutBody'
      responses:
        '200':
          description: Refresh thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login success
                  result:
                    $ref: '#/components/schemas/SuccessAuthenticate'
        '400':
          description: Nền tảng không hợp lệ
        '401':
          description: Token không hợp lệ
        '422':
          description: Sai các trường thông tin
...
components:
  schemas:
...
    LogoutBody:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresToken
...
    SuccessAuthenticate:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MCwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTY5MTQ2OTYyNn0.HTLX20cB7_z0c9c8FDg3MIx6RJEELHHlmJNZa94ku-o
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT